+++ source/intersection/intersect_cylinder.c
@@ -6,7 +6,7 @@
 /*   By: rcatini <rcatini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/22 11:11:27 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/03/24 15:57:34 by rcatini          ###   ########.fr       */
+/*   Updated: 2023/03/25 16:49:46 by rcatini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,8 +35,8 @@ t_intersection	intersect_cylinder(t_ray ray, t_object *cylinder)
 	if (abc_discri_calculating(&abc, &discriminant, ray))
 		return (ret);
 	ret.t.a = 2;
-	ty.a = (-abc.b - sqrtf(discriminant)) / (2 * abc.a);
-	ty.b = (-abc.b + sqrtf(discriminant)) / (2 * abc.a);
+	ty.a = (-abc.b - sqrt(discriminant)) / (2 * abc.a);
+	ty.b = (-abc.b + sqrt(discriminant)) / (2 * abc.a);
 	if (ty.a > ty.b)
 		swap(&ty.a, &ty.b);
 	ty.c = ray.point.y + ty.a * ray.vector.y;
@@ -51,12 +51,12 @@ t_intersection	intersect_cylinder(t_ray ray, t_object *cylinder)
 
 static int	abc_discri_calculating(t_3f *abc, double *discriminant, t_ray ray)
 {
-	abc->a = powf(ray.vector.x, 2) + powf(ray.vector.z, 2);
+	abc->a = pow(ray.vector.x, 2) + pow(ray.vector.z, 2);
 	if (equal_double(abc->a, 0))
 		return (1);
 	abc->b = 2 * ray.point.x * ray.vector.x + 2 * ray.point.z * ray.vector.z;
-	abc->c = powf(ray.point.x, 2) + powf(ray.point.z, 2) - 1;
-	*discriminant = powf(abc->b, 2) - 4 * abc->a * abc->c;
+	abc->c = pow(ray.point.x, 2) + pow(ray.point.z, 2) - 1;
+	*discriminant = pow(abc->b, 2) - 4 * abc->a * abc->c;
 	if (*discriminant < 0)
 		return (1);
 	return (0);
