+++ source/tuple/magnitude_vector.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   magnitude_vector.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rcatini <rcatini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/18 17:31:40 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/02/18 17:39:58 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/03/25 16:49:46 by rcatini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@ double	magnitude_vector(t_tuple *tuple)
 {
 	double	v;
 
-	v = sqrtf(powf(tuple->x, 2)
-			+ powf(tuple->y, 2)
-			+ powf(tuple->z, 2)
-			+ powf(tuple->w, 2));
+	v = sqrt(pow(tuple->x, 2)
+			+ pow(tuple->y, 2)
+			+ pow(tuple->z, 2)
+			+ pow(tuple->w, 2));
 	return (v);
 }
